First screen (Showing movies list) partially completed

+++ app/src/main/java/in/movies/niranjan/com/movieproj/models/Movie.java
@@ -28,7 +28,7 @@ public class Movie {
     public String overview;
 
     @SerializedName("popularity")
-    public Double popularity;
+    public float popularity;
 
     @SerializedName("poster_path")
     public String posterPath;
